backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
# psycopg2-binary==2.9.9  # Will use SQLite for development
Pillow==11.3.0
python-decouple==3.8
django-filter==25.2
# celery==5.3.4
# redis==5.0.1
# django-storages==1.14.2
# boto3==1.34.0
# whitenoise==6.6.0
# gunicorn==21.2.0
